requirements.txt
psycopg2-binary
SQLAlchemy>=2.0
pandas
python-dotenv
streamlit
